requirements.txt
blinker==1.9.0
click==8.3.1
dotenv==0.9.9
flash==1.0.3
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
python-dotenv==1.2.1
Werkzeug==3.1.3
